completely rewrite of nodewatcher2 due to the unbearable an unmaintainable pseudo-array code for emulating bash-arrays.
nodewatcher2 is now written in lua and can be configured by uci. It outputs json- and xml-files containing the nodedata. Optionally those files can
additionally be gzipped to save some traffic. Test showed gzipping costs nearly no time and reduces the size of the json- and xml-files by about 60%)

The nodewatcher2 divides its operation into two parts:

1) At first the lua-script collects all data it needs from the the system and puts it in a lua-table
2) The script now uses this lua-table to generate a abitrary set of output files by using it's defined and enabled output generators (xml, json,...)

The generated files then are put in /tmp/nodedata, which is symlinked into the /nodedata-url of the http-server of the gluon status-page

TODO:
- make nodewatcher-compat generator
- remove files whose generator was disabled via uci in the mean time